# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VectorVest Community Platform API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type ChannelCategory = typeof ChannelCategory[keyof typeof ChannelCategory];


export const ChannelCategory = {
  WELCOME: 'WELCOME',
  COMMUNITY: 'COMMUNITY',
  RESOURCES: 'RESOURCES',
} as const;

export interface Channel {
  id: string;
  name: string;
  description: string;
  category: ChannelCategory;
  writable: boolean;
  icon: string;
  unreadCount?: number;
}

export interface ReactionGroup {
  emoji: string;
  count: number;
  userReacted: boolean;
}

export interface Message {
  id: number;
  channelId: string;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  createdAt: string;
  reactions: ReactionGroup[];
  isPinned: boolean;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export interface ReactionInput {
  emoji: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  member: 'member',
} as const;

export type UserProfileMembershipStatus = typeof UserProfileMembershipStatus[keyof typeof UserProfileMembershipStatus];


export const UserProfileMembershipStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface UserProfile {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserProfileRole;
  membershipStatus: UserProfileMembershipStatus;
}

export interface UserProfileUpdate {
  /** @minLength 2 */
  username?: string;
}

export interface OnlineUser {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface OnlineUsersResponse {
  count: number;
  users: OnlineUser[];
}

export type WebinarStatus = typeof WebinarStatus[keyof typeof WebinarStatus];


export const WebinarStatus = {
  upcoming: 'upcoming',
  live: 'live',
  past: 'past',
} as const;

export interface Webinar {
  id: number;
  title: string;
  description: string;
  hostName: string;
  scheduledAt: string;
  status: WebinarStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  sessionNumber?: number | null;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  /** @nullable */
  location?: string | null;
  type?: string;
}

export type CourseLevel = typeof CourseLevel[keyof typeof CourseLevel];


export const CourseLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Course {
  id: number;
  title: string;
  description: string;
  instructor: string;
  /** @nullable */
  duration?: string | null;
  level: CourseLevel;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isLocked?: boolean;
}

export type MemberSummaryRole = typeof MemberSummaryRole[keyof typeof MemberSummaryRole];


export const MemberSummaryRole = {
  admin: 'admin',
  member: 'member',
} as const;

export type MemberSummaryMembershipStatus = typeof MemberSummaryMembershipStatus[keyof typeof MemberSummaryMembershipStatus];


export const MemberSummaryMembershipStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface MemberSummary {
  userId: string;
  username: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: MemberSummaryRole;
  membershipStatus: MemberSummaryMembershipStatus;
  /** @nullable */
  membershipMethod: string | null;
  /** @nullable */
  membershipVerifiedAt: string | null;
  createdAt: string;
}

export interface VerifyMemberInput {
  vectorVestMemberId?: string;
}

export interface RejectMemberInput {
  reason?: string;
}

export interface AllowlistEntry {
  id: number;
  email: string;
  /** @nullable */
  vectorVestMemberId: string | null;
  addedBy: string;
  addedAt: string;
  /** @nullable */
  notes: string | null;
}

export interface AllowlistEntryInput {
  email: string;
  vectorVestMemberId?: string;
  notes?: string;
}

export type GetChannelMessagesParams = {
/**
 * Cursor for pagination - get messages before this message ID
 */
before?: number;
limit?: number;
};

export type ListMembersParams = {
status?: ListMembersStatus;
};

export type ListMembersStatus = typeof ListMembersStatus[keyof typeof ListMembersStatus];


export const ListMembersStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;