# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: VectorVest Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: List all channels
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
        "401":
          description: Unauthorized
  /channels/{channelId}/messages:
    get:
      operationId: getChannelMessages
      tags: [messages]
      summary: Get messages for a channel
      security: [{ cookieAuth: [] }]
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          required: false
          schema:
            type: integer
          description: Cursor for pagination - get messages before this message ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Channel not found

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a channel
      security: [{ cookieAuth: [] }]
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Channel is read-only
...
  /channels/{channelId}/pinned:
    get:
      operationId: getPinnedMessage
      tags: [channels]
      summary: Get the pinned message for a channel
      security: [{ cookieAuth: [] }]
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pinned message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: No pinned message
...
  /admin/members:
    get:
      operationId: listMembers
      tags: [admin]
      summary: List members, optionally filtered by membership status
      security: [{ cookieAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, verified, rejected]
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberSummary"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/members/{userId}/verify:
    post:
      operationId: verifyMember
      tags: [admin]
      summary: Manually mark a member as verified VectorVest customer
      security: [{ cookieAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyMemberInput"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberSummary"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/members/{userId}/reject:
    post:
      operationId: rejectMember
      tags: [admin]
      summary: Manually reject a member's VectorVest membership claim
      security: [{ cookieAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectMemberInput"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberSummary"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/allowlist:
    get:
      operationId: listAllowlist
      tags: [admin]
      summary: List the VectorVest member email allowlist
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: Allowlist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllowlistEntry"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: addAllowlistEntry
      tags: [admin]
      summary: Add an email to the VectorVest member allowlist
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllowlistEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllowlistEntry"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/allowlist/{id}:
    delete:
      operationId: removeAllowlistEntry
      tags: [admin]
      summary: Remove an email from the VectorVest member allowlist
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    Channel:
      type: object
      required: [id, name, description, category, writable, icon]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [WELCOME, COMMUNITY, RESOURCES]
        writable:
          type: boolean
        icon:
          type: string
        unreadCount:
          type: integer
    Message:
      type: object
      required: [id, channelId, userId, username, content, createdAt, reactions, isPinned]
      properties:
        id:
          type: integer
        channelId:
          type: string
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionGroup"
        isPinned:
          type: boolean
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4000
    ReactionGroup:
      type: object
      required: [emoji, count, userReacted]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean
...
    MemberSummary:
      type: object
      required: [userId, username, email, role, membershipStatus, membershipMethod, membershipVerifiedAt, createdAt]
      properties:
        userId:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        membershipStatus:
          type: string
          enum: [pending, verified, rejected]
        membershipMethod:
          type: ["string", "null"]
        membershipVerifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    VerifyMemberInput:
      type: object
      properties:
        vectorVestMemberId:
          type: string
    RejectMemberInput:
      type: object
      properties:
        reason:
          type: string
    AllowlistEntry:
      type: object
      required: [id, email, vectorVestMemberId, addedBy, addedAt, notes]
      properties:
        id:
          type: integer
        email:
          type: string
        vectorVestMemberId:
          type: ["string", "null"]
        addedBy:
          type: string
        addedAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
    AllowlistEntryInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        vectorVestMemberId:
          type: string
        notes:
          type: string